+++ ExilenceNextApp/src/components/item-table/item-table-filter-subtotal/ItemTableFilterSubtotal.tsx
@@ -37,7 +37,7 @@ const ItemTableFilterSubtotal = ({ array }: ItemTableFilterSubtotalProps) => {
 
   return (
     <>
-      {uiStateStore?.itemTableFilterText !== '' && (
+      {(uiStateStore?.itemTableFilterText !== '' || uiStateStore?.showItemTableFilter) && (
         <Paper className={clsx(classes.paper)}>
           {t('label.filter_total')}
           <span style={{ color: rarityColors.currency }}>